Reproducer for device_adjacent_difference hipGraph unit test bug
The device_adjacent_difference test fails intermittently when
run using hipGraphs on gfx1101. The issue appears to be related
to the hipMemset calls the are made into the default stream
just before stream capture is started. If I insert a call to
hipStreamSynchronize(0) between the hipMemset calls and the
beginning of the graph capture, the test consistently passes.

I believe this behaviour should not be occurring because calls
submitted to the default stream are supposed to be run synchronously.