Workaround for shared memory write ordering compiler issue
This change implements a workaround to fix block_histogram while we wait for a compiler fix.

The block_histogram_sort algorithm contains a loop that initializes values in a shared memory struct.
Then it runs the flag_heads() algorithm using that struct.
It appear that in some cases, the optimizer is reordering things such that shared memory writes in
flag_heads() run before the loop initialization code completes.

Changing the loop condition and removing the inner "if" statement seems to prevent the optimizer from
doing this. From an algorithmic standpoint, this should be functionally equivalent to the way the code
was before the change.