Set default accumulator type for device_scan to output type
Currently, we set the default accumulator type for device-wide scans
to the type of the init value, or the input type. This can result in
situations where the accumulator overflows if the init value type is
smaller than the output type.

This change modifies the default accumulator type, setting it to the
output type.