Add hipGetLastError calls to clear existing errors
The behaviour of hipGetLastError will be changing in an
upcoming HIP release. Currently, the error that's reported
is cleared on each HIP API call. This means that
hipGetLastError reports any error that occurred during the last
HIP API call (in the current host thread).

Moving forward, the error that's reported will only be cleared
on each call to hipGetLastError. This means that hipGetLastError
will report any error that has occurred since the last call to
hipGetError (in the current host thread).

Some of our tests rely on observing a return value of
hipErrorOutOfMemory from hipMalloc when an allocation is too
large for a given GPU architecture's memory system. While this
will still work with the future behaviour, it will cause
subsequent calls to hipGetError to also report this error.

This change fixes these tests by calling hipGetLastError before
sections of code we want to detect errors in, so that any previously
recorded error is cleared. This ensures that when we call hipGetLastError
again after the code sections of interest complete, it only reports
errors from within the sections of interest.

This change also adds error-clearing hipGetLastError calls to other
locations (besides the tests mentioned above) where hipGetLastError
is called. This is mainly to guard against user code that makes
HIP API calls which set an error between rocPRIM calls.

More specifically, this change addes code to clear errors before:
- existing calls to hipGetLastError
- calls to the ROCPRIM_DETAIL_HIP_SYNC_AND_RETURN_ON_ERROR macro,
which calls hipGetLastError