Support getting device ID from stream hipStreamLegacy (#687)
rocprim::get_device_from_stream must retreive the device ID
in different ways depending on whether a user-created stream
or a built-in stream type is being used.

This change modifies the check for built-in stream types so that
it includes hipStreamLegacy. This prevents the function from
calling hipGetStreamDeviceId(hipStreamLegacy), which currently
causes a segmentation fault.